proto/agent: add delegation API
Add proto definitios for new delegation API. This API includes two
rpcs: FetchX509SVIDs and FetchX509Bundles.

The FetchX509SVIDs rpc allows a privileged client to get X509-SVIDs
for a given set of workloads. This method uses a bidirectional gRPC
stream. The server to client flow is used to send SVID updates and
the client to server one is used as a control mechanism to add /
remove workloads from the watched set.

The client can add workloads to the watched set by specifying a list of
workload's selectors. X509-SVIDs for registration entries matching
*all* the passed selectors are returned.

The FetchX509Bundles rpc streams all bundles (the federated and the
trust domain one) to the client.

Signed-off-by: Mauricio Vásquez <mauricio@accuknox.com>
Signed-off-by: Mauricio Vásquez <mauricio@kinvolk.io>