Simplify the Logger API interface.
Remove the LoggerMask type and support for it in GetLogger and
ListLogger RPCs.  The Logger type is only a string and two enums (ints)
so making the enums optional doesn't add up to significant savings
compared to the message complexity it would add.

Remove the CountLoggers RPC.  Users of the API can use the ListLoggers
RPC and then count the returned entries.  This simplifies the interface
and removes any count mismatches between a call to CountLoggers and
ListLoggers.

Remove the ListLoggerRequest sub-message Filter, promoting the name and
with_subloggers options into the ListLoggerRequest.  This simplifies the
request by not burdening it with a sub-message.

Fixed all the comments, they previously had references to the server's
agent API, the inspiration of the original API structure.

Signed-off-by: Edwin Buck <edwbuck@gmail.com>